+++ scripts/getJwtToken.ts
@@ -10,9 +10,9 @@ import { createTestConnection } from '../tests/utils/testConnection'
 async function getAdminUserToken() {
     await createTestConnection()
     // email taken from ADMIN_EMAILS list
-    let user = await User.findOne({ email: 'sandy@calmid.com' })
+    let user = await User.findOne({ email: 'sandy@kidsloop.live' })
     if (user == undefined) {
-        user = await createUser({ email: 'sandy@calmid.com' }).save()
+        user = await createUser({ email: 'sandy@kidsloop.live' }).save()
     }
 
     const tokenInfo = {