+++ src/resolvers/user.ts
@@ -3,14 +3,9 @@ import { OrganizationMembership } from '../entities/organizationMembership'
 import { Role } from '../entities/role'
 import { User } from '../entities/user'
 import { Context } from '../main'
-import {
-    CoreUserConnectionNode,
-    mapUserToUserConnectionNode,
-} from '../pagination/usersConnection'
+import { mapUserToUserConnectionNode } from '../pagination/usersConnection'
 import { PermissionName } from '../permissions/permissionNames'
-import { UserPermissions } from '../permissions/userPermissions'
-import { APIError, APIErrorCollection } from '../types/errors/apiError'
-import { customErrors } from '../types/errors/customError'
+import { APIError } from '../types/errors/apiError'
 import {
     AddOrganizationRolesToUserInput,
     AddSchoolRolesToUserInput,
@@ -27,12 +22,15 @@ import {
 import {
     createEntityAPIError,
     createExistentEntityAttributesAPIError,
-    createInputLengthAPIError,
 } from '../utils/resolvers/errors'
-import { getMap, SchoolMembershipMap } from '../utils/resolvers/entityMaps'
-import { config } from '../config/config'
+import {
+    getMap,
+    OrganizationMembershipMap,
+    SchoolMembershipMap,
+} from '../utils/resolvers/entityMaps'
 import {
     AddMutation,
+    AddRemoveMutation,
     CreateMutation,
     EntityMap,
     filterInvalidInputs,
@@ -63,6 +61,7 @@ import {
     cleanUpdateUserInput,
     ConflictingUserKey,
     createUserInputToConflictingUserKey,
+    OrganizationMembershipKey,
     updateUserInputToConflictingUserKey,
 } from '../utils/resolvers/user'
 
@@ -72,6 +71,19 @@ interface CreateUsersEntityMap extends EntityMap<User> {
 
 interface UpdateUsersEntityMap extends CreateUsersEntityMap {
     mainEntity: Map<string, User>
+}
+
+interface AddOrganizationRolesToUsersEntityMap
+    extends RemoveOrganizationRolesFromUsersEntityMap {
+    orgRoles: Map<string, Role[]>
+}
+
+interface RemoveOrganizationRolesFromUsersEntityMap extends EntityMap<User> {
+    mainEntity: Map<string, User>
+    organizations: Map<string, Organization>
+    roles: Map<string, Role>
+    memberships: OrganizationMembershipMap
+    membershipRoles: ObjMap<{ organizationId: string; userId: string }, Role[]>
 }
 
 export interface AddSchoolRolesToUsersEntityMap
@@ -87,6 +99,18 @@ interface RemoveSchoolRolesFromUsersEntityMap extends EntityMap<User> {
     memberships: SchoolMembershipMap
     membershipRoles: ObjMap<{ schoolId: string; userId: string }, Role[]>
 }
+
+type AddRemoveOrganizationRolesInputTypeName =
+    | 'AddOrganizationRolesToUserInput'
+    | 'RemoveOrganizationRolesFromUserInput'
+
+export type AddRemoveOrganizationRolesInputType =
+    | AddOrganizationRolesToUserInput
+    | RemoveOrganizationRolesFromUserInput
+
+export type AddRemoveOrganizationRolesEntityMap =
+    | AddOrganizationRolesToUsersEntityMap
+    | RemoveOrganizationRolesFromUsersEntityMap
 
 export class AddSchoolRolesToUsers extends AddMutation<
     User,
@@ -324,123 +348,282 @@ export class AddSchoolRolesToUsers extends AddMutation<
     }
 }
 
-export function addOrganizationRolesToUsers(
-    args: { input: AddOrganizationRolesToUserInput[] },
-    context: Pick<Context, 'permissions'>
-): Promise<UsersMutationResult> {
-    return modifyOrganizationRoles(
-        args,
-        context.permissions,
-        (currentRoles: Role[], rolesToAdd: Role[]): Role[] => {
-            currentRoles.push(...rolesToAdd)
-            return [...new Set(currentRoles)]
-        }
-    )
-}
+abstract class AddRemoveOrganizationRoles<
+    InputType extends AddRemoveOrganizationRolesInputType,
+    EntityMapType extends AddRemoveOrganizationRolesEntityMap
+> extends AddRemoveMutation<
+    User,
+    InputType,
+    UsersMutationResult,
+    EntityMapType,
+    OrganizationMembership
+> {
+    protected EntityType = User
+    protected abstract inputTypeName: AddRemoveOrganizationRolesInputTypeName
+    protected mainEntityIds: string[]
+    protected output: UsersMutationResult = { users: [] }
 
-export function removeOrganizationRolesFromUsers(
-    args: { input: RemoveOrganizationRolesFromUserInput[] },
-    context: Pick<Context, 'permissions'>
-): Promise<{ users: CoreUserConnectionNode[] }> {
-    return modifyOrganizationRoles(
-        args,
-        context.permissions,
-        (currentRoles: Role[], rolesToRemove: Role[]): Role[] =>
-            currentRoles.filter(
-                (cr) => !rolesToRemove.find((rtr) => cr.role_id === rtr.role_id)
+    constructor(input: InputType[], permissions: Context['permissions']) {
+        super(input, permissions)
+        this.mainEntityIds = input.map((i) => i.userId)
+    }
+
+    async generateEntityMaps(input: InputType[]): Promise<EntityMapType> {
+        let organizationIds: string[] = []
+        let roleIds: string[] = []
+        let userIds: string[] = []
+
+        input.forEach((i) => {
+            organizationIds.push(i.organizationId)
+            roleIds.push(...i.roleIds)
+            userIds.push(i.userId)
+        })
+
+        organizationIds = [...new Set(organizationIds)]
+        roleIds = [...new Set(roleIds)]
+        userIds = [...new Set(userIds)]
+
+        const userMap = getMap.user(userIds)
+        const roleMap = getMap.role(roleIds, ['organization'])
+        const organizationMap = await getMap.organization(organizationIds)
+        const organizationMapKeys = [...organizationMap.keys()]
+        const orgRoles = new Map<string, Role[]>(
+            Array.from(organizationMapKeys, (k) => [k, []])
+        )
+
+        const isAdding =
+            this.inputTypeName === 'AddOrganizationRolesToUserInput'
+        if (isAdding) {
+            const allRolesFromOrgs = await Role.find({
+                // don't query system roles, as these are valid for all organizations anyway
+                where: {
+                    system_role: false,
+                    organization: {
+                        organization_id: In(organizationMapKeys),
+                    },
+                },
+                join: {
+                    alias: 'Role',
+                    leftJoinAndSelect: {
+                        organization: 'Role.organization',
+                    },
+                },
+            })
+
+            const rolesOrgs = await Promise.all(
+                allRolesFromOrgs.map((r) => r.organization!)
             )
-    )
-}
 
-async function modifyOrganizationRoles(
-    args: {
-        input:
-            | AddOrganizationRolesToUserInput[]
-            | RemoveOrganizationRolesFromUserInput[]
-    },
-    permissions: UserPermissions,
-    roleModificationFn: (currentRoles: Role[], roleChanges: Role[]) => Role[]
-) {
-    // Initial validations
-    if (args.input.length === 0) throw createInputLengthAPIError('User', 'min')
-    if (args.input.length > config.limits.MUTATION_MAX_INPUT_ARRAY_SIZE)
-        throw createInputLengthAPIError('User', 'max')
-    await permissions.rejectIfNotAllowed(
-        { organization_ids: args.input.map((i) => i.organizationId) },
-        PermissionName.edit_users_40330
-    )
+            for (const [i, role] of allRolesFromOrgs.entries()) {
+                const roleOrgId = rolesOrgs[i].organization_id
+                const roles = orgRoles.get(roleOrgId) ?? []
+                roles.push(role)
+                orgRoles.set(roleOrgId, roles)
+            }
+        }
 
-    // Preloading
-    const orgIds = args.input.map((val) => val.organizationId)
-    const userIds = args.input.map((val) => val.userId)
-    const orgMapPromise = getMap.organization(orgIds)
-    const userMapPromise = getMap.user(userIds)
-    const roleMapPromise = getMap.role(args.input.flatMap((val) => val.roleIds))
-    const membershipMapPromise = getMap.membership.organization(
-        orgIds,
-        userIds,
-        ['roles']
-    )
-    const preloadedOrganizations = await orgMapPromise
-    const preloadedUsers = await userMapPromise
-    const preloadedRoles = await roleMapPromise
-    const preloadedMemberships = await membershipMapPromise
+        const membershipMap = await getMap.membership.organization(
+            organizationIds,
+            userIds,
+            ['roles']
+        )
 
-    // Process inputs
-    const errors: APIError[] = []
-    const memberships: OrganizationMembership[] = []
-    const output: CoreUserConnectionNode[] = []
-    for (const [index, subArgs] of args.input.entries()) {
-        const { organizationId, userId, roleIds } = subArgs
+        // Make map for memberships' roles (does not need to query db)
+        const membershipRoles = new ObjMap<OrganizationMembershipKey, Role[]>()
+        for (const [key, membership] of membershipMap.entries()) {
+            if (membership.roles) {
+                // eslint-disable-next-line no-await-in-loop
+                membershipRoles.set(key, await membership.roles)
+            }
+        }
+
+        const maps = {
+            mainEntity: await userMap,
+            organizations: organizationMap,
+            roles: await roleMap,
+            memberships: membershipMap,
+            membershipRoles,
+        }
+
+        if (isAdding) {
+            Object.assign(maps, { orgRoles })
+        }
+
+        return maps as EntityMapType
+    }
+
+    async authorize(input: InputType[]): Promise<void> {
+        const organizationIds = input.map((i) => i.organizationId)
+        return this.permissions.rejectIfNotAllowed(
+            { organization_ids: organizationIds },
+            PermissionName.edit_users_40330
+        )
+    }
+
+    validationOverAllInputs(
+        inputs: InputType[]
+    ): {
+        validInputs: { index: number; input: InputType }[]
+        apiErrors: APIError[]
+    } {
+        const failedDuplicateUsers = validateNoDuplicateAttribute(
+            inputs.map((i) => {
+                return { entityId: i.organizationId, attributeValue: i.userId }
+            }),
+            'Organization',
+            'userId'
+        )
+
+        const failedRoles = validateSubItemsLengthAndNoDuplicates(
+            inputs,
+            this.inputTypeName,
+            'roleIds'
+        )
+
+        return filterInvalidInputs(inputs, [
+            failedDuplicateUsers,
+            ...failedRoles,
+        ])
+    }
+
+    validate(
+        index: number,
+        _currentUser: User | undefined,
+        currentInput: InputType,
+        maps: EntityMapType
+    ): APIError[] {
+        const { userId, organizationId, roleIds } = currentInput
+        const errors: APIError[] = []
 
         const orgs = flagNonExistent(
             Organization,
             index,
             [organizationId],
-            preloadedOrganizations
+            maps.organizations
         )
-        const roles = flagNonExistent(Role, index, roleIds, preloadedRoles)
-        const users = flagNonExistent(User, index, [userId], preloadedUsers)
-        errors.push(...orgs.errors, ...roles.errors, ...users.errors)
 
-        if (orgs.errors.length || users.errors.length) continue
-        const dbMemberships = flagNonExistentOrganizationMembership(
+        const users = flagNonExistent(User, index, [userId], maps.mainEntity)
+        const roles = flagNonExistent(Role, index, roleIds, maps.roles)
+        errors.push(...orgs.errors, ...users.errors, ...roles.errors)
+
+        if (users.errors.length || orgs.errors.length) return errors
+
+        const memberships = flagNonExistentOrganizationMembership(
             index,
             organizationId,
             [userId],
-            preloadedMemberships
+            maps.memberships
         )
-        if (dbMemberships.errors) errors.push(...dbMemberships.errors)
 
-        for (const dbMembership of dbMemberships.values) {
-            if (errors.length > 0) continue
-            // eslint-disable-next-line no-await-in-loop
-            const dbMembershipRoles = (await dbMembership.roles) || [] // already fetched
-            dbMembership.roles = Promise.resolve(
-                roleModificationFn(dbMembershipRoles, roles.values)
+        errors.push(...memberships.errors)
+
+        if (memberships.errors.length) return errors
+
+        const isAdding =
+            this.inputTypeName === 'AddOrganizationRolesToUserInput'
+
+        const roleChildMethod = isAdding
+            ? flagExistentChild
+            : flagNonExistentChild
+
+        const existingRoles = maps.membershipRoles.get({
+            organizationId,
+            userId,
+        })!
+
+        const roleInMembershipErrors = roleChildMethod(
+            Organization,
+            Role,
+            index,
+            organizationId,
+            roles.values.map((r) => r.role_id),
+            new Set(existingRoles.map((r) => r.role_id))
+        )
+
+        errors.push(...roleInMembershipErrors)
+
+        if (isAdding) {
+            const mapsToAdd = maps as AddOrganizationRolesToUsersEntityMap
+            const orgRoles = mapsToAdd.orgRoles.get(organizationId)!
+            const roleInOrgErrors = flagNonExistentChild(
+                Organization,
+                Role,
+                index,
+                organizationId,
+                roles.values
+                    .filter((r) => !r.system_role)
+                    .map((r) => r.role_id),
+                new Set(orgRoles.map((r) => r.role_id))
             )
-            memberships.push(dbMembership)
+
+            errors.push(...roleInOrgErrors)
         }
 
-        // Build output
-        if (errors.length > 0) continue
-        output.push(mapUserToUserConnectionNode(users.values[0]))
+        return errors
     }
 
-    if (errors.length > 0) throw new APIErrorCollection(errors)
-    try {
-        await getManager().save(memberships)
-    } catch (e) {
-        const message = e instanceof Error ? e.message : 'Unknown Error'
-        throw new APIError({
-            code: customErrors.database_save_error.code,
-            message: customErrors.database_save_error.message,
-            variables: [message],
-            entity: 'User',
-        })
+    protected process(
+        currentInput: InputType,
+        maps: EntityMapType,
+        index: number
+    ): ProcessedResult<User, OrganizationMembership> {
+        const currentEntity = maps.mainEntity.get(this.mainEntityIds[index])!
+        const { userId, organizationId, roleIds } = currentInput
+        const dbMembership = maps.memberships.get({ organizationId, userId })!
+        const dbMembershipRoles = maps.membershipRoles.get({
+            organizationId,
+            userId,
+        })!
+
+        const isAdding =
+            this.inputTypeName === 'AddOrganizationRolesToUserInput'
+
+        let newRoles: Role[] = []
+        if (isAdding) {
+            newRoles = [
+                ...dbMembershipRoles,
+                ...roleIds.map((rId) => maps.roles.get(rId)!),
+            ]
+        } else {
+            const roleIdsSet = new Set(roleIds)
+            newRoles = dbMembershipRoles.filter(
+                (dmr) => !roleIdsSet.has(dmr.role_id)
+            )
+        }
+
+        dbMembership.roles = Promise.resolve(newRoles)
+        return { outputEntity: currentEntity, modifiedEntity: [dbMembership] }
     }
 
-    return { users: output }
+    protected async applyToDatabase(
+        results: ProcessedResult<User, OrganizationMembership>[]
+    ): Promise<void> {
+        const saveEntities = results
+            .flatMap((r) => r.modifiedEntity)
+            .filter((r): r is OrganizationMembership => r !== undefined)
+
+        await getManager().save(saveEntities)
+    }
+
+    protected async buildOutput(currentUser: User): Promise<void> {
+        this.output.users.push(mapUserToUserConnectionNode(currentUser))
+    }
+}
+
+export class AddOrganizationRolesToUsers extends AddRemoveOrganizationRoles<
+    AddOrganizationRolesToUserInput,
+    AddOrganizationRolesToUsersEntityMap
+> {
+    protected inputTypeName: AddRemoveOrganizationRolesInputTypeName =
+        'AddOrganizationRolesToUserInput'
+}
+
+export class RemoveOrganizationRolesFromUsers extends AddRemoveOrganizationRoles<
+    RemoveOrganizationRolesFromUserInput,
+    RemoveOrganizationRolesFromUsersEntityMap
+> {
+    protected inputTypeName: AddRemoveOrganizationRolesInputTypeName =
+        'RemoveOrganizationRolesFromUserInput'
 }
 
 export class RemoveSchoolRolesFromUsers extends RemoveMutation<