+++ src/schemas/class.ts
@@ -22,6 +22,8 @@ import {
     AddTeachersToClasses,
     RemoveTeachersFromClasses,
     SetAcademicTermsOfClasses,
+    moveUsersToClass,
+    moveUsersTypeToClass,
 } from '../resolvers/class'
 import { IChildConnectionDataloaderKey } from '../loaders/childConnectionLoader'
 import { Subject } from '../entities/subject'
@@ -62,6 +64,12 @@ const typeDefs = gql`
         setAcademicTermsOfClasses(
             input: [SetAcademicTermOfClassInput!]!
         ): ClassesMutationResult
+        moveStudentsToClass(
+            input: MoveUsersToClassInput
+        ): MoveUsersToClassMutationResult
+        moveTeachersToClass(
+            input: MoveUsersToClassInput
+        ): MoveUsersToClassMutationResult
     }
 
     # pagination extension types start here
@@ -88,6 +96,17 @@ const typeDefs = gql`
         order: SortOrder!
     }
 
+    input MoveUsersToClassInput {
+        fromClassId: ID!
+        toClassId: ID!
+        userIds: [ID!]!
+    }
+
+    type MoveUsersToClassMutationResult {
+        fromClass: ClassConnectionNode!
+        toClass: ClassConnectionNode!
+    }
+
     input ClassFilter {
         id: UUIDFilter
         name: StringFilter
@@ -102,6 +121,7 @@ const typeDefs = gql`
         schoolId: UUIDExclusiveFilter
         gradeId: UUIDFilter
         subjectId: UUIDFilter
+        academicTermId: UUIDExclusiveFilter
 
         #connections - extra filters
         studentId: UUIDFilter
@@ -581,6 +601,18 @@ export default function getDefault(
                     mutate(RemoveTeachersFromClasses, args, ctx.permissions),
                 setAcademicTermsOfClasses: (_parent, args, ctx) =>
                     mutate(SetAcademicTermsOfClasses, args, ctx.permissions),
+                moveStudentsToClass: (_parent, args, ctx) =>
+                    moveUsersToClass(
+                        ctx,
+                        args.input,
+                        moveUsersTypeToClass.students
+                    ),
+                moveTeachersToClass: (_parent, args, ctx) =>
+                    moveUsersToClass(
+                        ctx,
+                        args.input,
+                        moveUsersTypeToClass.teachers
+                    ),
             },
             Query: {
                 class: (_parent, args, ctx, _info) => model.getClass(args, ctx),