+++ tests/integration/utils/csv/user.test.ts
@@ -1760,11 +1760,7 @@ describe('processUserFromCSVRow', async () => {
         })
 
         context('and the role is neither student nor teacher related', () => {
-            beforeEach(() =>
-                createRoleForUser('notAStudentOrTeacher', [
-                    PermissionName.add_learning_outcome_to_content_485,
-                ])
-            )
+            beforeEach(() => createRoleForUser('notAStudentOrTeacher', []))
 
             it('raises an UNAUTHORIZED_UPLOAD_CHILD_ENTITY error against the Role column', async () => {
                 const rowErrors = await processRow()
@@ -1782,7 +1778,6 @@ describe('processUserFromCSVRow', async () => {
         context('and the role is student related', () => {
             beforeEach(() =>
                 createRoleForUser('Pupil', [
-                    PermissionName.add_learning_outcome_to_content_485,
                     PermissionName.attend_live_class_as_a_student_187,
                 ])
             )
@@ -1794,7 +1789,6 @@ describe('processUserFromCSVRow', async () => {
         context('and the role is teacher related', () => {
             beforeEach(() =>
                 createRoleForUser('Master', [
-                    PermissionName.add_learning_outcome_to_content_485,
                     PermissionName.attend_live_class_as_a_teacher_186,
                 ])
             )
@@ -1806,7 +1800,6 @@ describe('processUserFromCSVRow', async () => {
         context('and the role is both student and teacher related', () => {
             beforeEach(() =>
                 createRoleForUser('MasterAndPupil', [
-                    PermissionName.add_learning_outcome_to_content_485,
                     PermissionName.attend_live_class_as_a_teacher_186,
                     PermissionName.attend_live_class_as_a_student_187,
                 ])
@@ -2092,7 +2085,6 @@ describe('processUserFromCSVRow', async () => {
                 () => {
                     beforeEach(() =>
                         createRoleForUser('Pupil', [
-                            PermissionName.add_learning_outcome_to_content_485,
                             PermissionName.attend_live_class_as_a_student_187,
                         ])
                     )